+++ src/main/java/demo/codeexample/auth/AuthController.java
@@ -3,7 +3,7 @@ package demo.codeexample.auth;
 import demo.codeexample.exception.UnauthorizedException;
 import demo.codeexample.security.JwtService;
 import demo.codeexample.user.User;
-import demo.codeexample.user.UserRepository;
+import demo.codeexample.user.doman.UserRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
