+++ src/main/java/demo/codeexample/security/OAuth2LoginSuccessHandler.java
@@ -2,6 +2,7 @@ package demo.codeexample.security;
 
 import demo.codeexample.user.doman.Role;
 import demo.codeexample.user.*;
+import demo.codeexample.user.doman.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.Authentication;
